:bug: Assigned dynamic to whole WritePage, not on TinyMCE
